new-borders - deltaVal is BigDecimal

+++ core/src/main/java/com/bitplay/api/controller/BordersEndpoint.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -44,26 +45,26 @@ public class BordersEndpoint {
     public BorderParams setDefaultBorders() {
         final List<BorderTable> borders = new ArrayList<>();
         final List<BorderItem> borderBtmClose = new ArrayList<>();
-        borderBtmClose.add(new BorderItem(1, -21, 500, 500));
-        borderBtmClose.add(new BorderItem(2, -15, 300, 350));
-        borderBtmClose.add(new BorderItem(3, -10, 200, 250));
-        borderBtmClose.add(new BorderItem(4, -5, 100, 100));
+        borderBtmClose.add(new BorderItem(1, BigDecimal.valueOf(-21), 500, 500));
+        borderBtmClose.add(new BorderItem(2, BigDecimal.valueOf(-15), 300, 350));
+        borderBtmClose.add(new BorderItem(3, BigDecimal.valueOf(-10), 200, 250));
+        borderBtmClose.add(new BorderItem(4, BigDecimal.valueOf(-5), 100, 100));
         borders.add(new BorderTable("b_br_close", borderBtmClose));
         final List<BorderItem> borderBtmOpen = new ArrayList<>();
-        borderBtmOpen.add(new BorderItem(1, 20, 100, 100));
-        borderBtmOpen.add(new BorderItem(2, 30, 250, 250));
-        borderBtmOpen.add(new BorderItem(3, 35, 350, 350));
+        borderBtmOpen.add(new BorderItem(1, BigDecimal.valueOf(20), 100, 100));
+        borderBtmOpen.add(new BorderItem(2, BigDecimal.valueOf(30), 250, 250));
+        borderBtmOpen.add(new BorderItem(3, BigDecimal.valueOf(35), 350, 350));
         borders.add(new BorderTable("b_br_open", borderBtmOpen));
         final List<BorderItem> borderOkexClose = new ArrayList<>();
-        borderOkexClose.add(new BorderItem(1, -21, 500, 500));
-        borderOkexClose.add(new BorderItem(2, -15, 300, 350));
-        borderOkexClose.add(new BorderItem(3, -10, 200, 250));
-        borderOkexClose.add(new BorderItem(4, -5, 100, 100));
+        borderOkexClose.add(new BorderItem(1, BigDecimal.valueOf(-21), 500, 500));
+        borderOkexClose.add(new BorderItem(2, BigDecimal.valueOf(-15), 300, 350));
+        borderOkexClose.add(new BorderItem(3, BigDecimal.valueOf(-10), 200, 250));
+        borderOkexClose.add(new BorderItem(4, BigDecimal.valueOf(-5), 100, 100));
         borders.add(new BorderTable("o_br_close", borderOkexClose));
         final List<BorderItem> borderOkexOpen = new ArrayList<>();
-        borderOkexOpen.add(new BorderItem(1, 20, 100, 100));
-        borderOkexOpen.add(new BorderItem(2, 30, 250, 250));
-        borderOkexOpen.add(new BorderItem(3, 35, 350, 350));
+        borderOkexOpen.add(new BorderItem(1, BigDecimal.valueOf(20), 100, 100));
+        borderOkexOpen.add(new BorderItem(2, BigDecimal.valueOf(30), 250, 250));
+        borderOkexOpen.add(new BorderItem(3, BigDecimal.valueOf(35), 350, 350));
         borders.add(new BorderTable("o_br_open", borderOkexOpen));
 
         final BorderParams borderParams = new BorderParams(BorderParams.Ver.V2, new BordersV1(), new BordersV2(borders));
